Assesment 3 Zakat Fitrah

+++ app/src/main/java/org/d3if2103/mitibleday/db/ZakatEntity.kt
@@ -8,6 +8,6 @@ data class ZakatEntity(
     @PrimaryKey(autoGenerate = true)
     var id: Long = 0L,
     var tanggal: Long = System.currentTimeMillis(),
-    var hargaBeras: Double,
-    var jiwa: Double
+    var hargaBeras: Float,
+    var jiwa: Float
 )
